refactor(Gateways): merge GatewayStorage into Gateway
