+++ src/lib/settings/gateway/Gateway.ts
@@ -1,10 +1,28 @@
 import { RequestHandler, IdKeyed } from '@klasa/request-handler';
-import { GatewayStorage, GatewayStorageOptions } from './GatewayStorage';
 import { Settings } from '../Settings';
 import { Cache } from '@klasa/cache';
-import type { Client } from '@klasa/core';
+import { Schema } from '../schema/Schema';
 
-export class Gateway extends GatewayStorage {
+import type { Client } from '@klasa/core';
+import type { Provider } from 'klasa';
+import type { SchemaEntryJson } from '../schema/SchemaEntry';
+
+export class Gateway {
+
+	/**
+	 * The client this gateway was created with.
+	 */
+	public readonly client: Client;
+
+	/**
+	 * The name of this gateway.
+	 */
+	public readonly name: string;
+
+	/**
+	 * The schema for this gateway.
+	 */
+	public readonly schema: Schema;
 
 	/**
 	 * The cached entries for this Gateway or the external datastore to get the settings from.
@@ -18,8 +36,21 @@ export class Gateway extends GatewayStorage {
 	 */
 	public readonly requestHandler: RequestHandler<string, IdKeyed<string>>;
 
-	public constructor(client: Client, name: string, options: GatewayStorageOptions = {}) {
-		super(client, name, options);
+	/**
+	 * Whether or not this gateway has been initialized.
+	 */
+	public ready = false;
+
+	/**
+	 * The provider's name that manages this gateway.
+	 */
+	private readonly _provider: string;
+
+	public constructor(client: Client, name: string, options: GatewayOptions = {}) {
+		this.client = client;
+		this.name = name;
+		this.schema = options.schema || new Schema();
+		this._provider = options.provider || client.options.providers.default || '';
 		this.cache = (this.name in this.client) && (this.client[this.name as keyof Client] instanceof Map) ?
 			this.client[this.name as keyof Client] as Map<string, ProxyMapEntry> :
 			new Cache<string, ProxyMapEntry>();
@@ -90,6 +121,49 @@ export class Gateway extends GatewayStorage {
 	}
 
 	/**
+	 * The provider that manages this gateway's persistent data.
+	 */
+	public get provider(): Provider | null {
+		return this.client.providers.get(this._provider) ?? null;
+	}
+
+	/**
+	 * Initializes the gateway.
+	 */
+	public async init(): Promise<void> {
+		// Gateways must not initialize twice.
+		if (this.ready) throw new Error(`The gateway "${this.name}" has already been initialized.`);
+
+		// Check the provider's existence.
+		const { provider } = this;
+		if (provider === null) throw new Error(`The gateway "${this.name}" could not find the provider "${this._provider}".`);
+		this.ready = true;
+
+		const errors = [...this._initializeSchemaEntries(this.schema)];
+		if (errors.length) throw new Error(`[SCHEMA] There is an error with your schema.\n${errors.join('\n')}`);
+
+		// Initialize the defaults
+		// eslint-disable-next-line dot-notation
+		this.schema.defaults['_init']();
+
+		// Init the table
+		const hasTable = await provider.hasTable(this.name);
+		if (!hasTable) await provider.createTable(this.name);
+
+		// Add any missing columns (NoSQL providers return empty array)
+		const columns = await provider.getColumns(this.name);
+		if (columns.length) {
+			const promises = [];
+			for (const entry of this.schema.values()) {
+				if (!columns.includes(entry.key)) promises.push(provider.addColumn(this.name, entry));
+			}
+			await Promise.all(promises);
+		}
+
+		await this.sync();
+	}
+
+	/**
 	 * Runs a synchronization task for the gateway.
 	 */
 	public async sync(): Promise<this> {
@@ -97,6 +171,48 @@ export class Gateway extends GatewayStorage {
 		return this;
 	}
 
+	/**
+	 * Get A JSON object containing the schema and the options.
+	 */
+	public toJSON(): GatewayJson {
+		return {
+			name: this.name,
+			provider: this._provider,
+			schema: this.schema.toJSON()
+		};
+	}
+
+	private *_initializeSchemaEntries(schema: Schema): IterableIterator<string> {
+		// Iterate over all the schema's values
+		for (const value of schema.values()) {
+			// Set the client and check if it is valid, afterwards freeze,
+			// otherwise delete it from its parent and yield error message
+			value.client = this.client;
+			try {
+				value._check();
+				Object.freeze(value);
+			} catch (error) {
+				// If errored, delete the key from its parent
+				value.parent.delete(value.key);
+				yield error.message;
+			}
+		}
+
+		// Set the schema as ready
+		schema.ready = true;
+	}
+
+}
+
+export interface GatewayOptions {
+	schema?: Schema;
+	provider?: string;
+}
+
+export interface GatewayJson {
+	name: string;
+	provider: string;
+	schema: Record<string, SchemaEntryJson>;
 }
 
 export interface ProxyMapEntry {
