+++ src/lib/settings/gateway/GatewayStore.ts
@@ -1,9 +1,9 @@
 import { Cache } from '@klasa/cache';
 
-import type { GatewayStorage, GatewayStorageJson } from './GatewayStorage';
 import type { Client } from '@klasa/core';
+import type { Gateway, GatewayJson } from './Gateway';
 
-export class GatewayDriver extends Cache<string, GatewayStorage> {
+export class GatewayStore extends Cache<string, Gateway> {
 
 	/**
 	 * The client this GatewayDriver was created with.
@@ -39,9 +39,9 @@ export class GatewayDriver extends Cache<string, GatewayStorage> {
 	 * // register calls can be chained
 	 * client
 	 *     .register(new Gateway(client, 'channels'))
-	 *     .register(new GatewayStorage(client, 'moderations', { provider: 'postgres' }));
+	 *     .register(new Gateway(client, 'moderations', { provider: 'postgres' }));
 	 */
-	public register(gateway: GatewayStorage): this {
+	public register(gateway: Gateway): this {
 		this.set(gateway.name, gateway);
 		return this;
 	}
@@ -62,4 +62,4 @@ export class GatewayDriver extends Cache<string, GatewayStorage> {
 
 }
 
-export type GatewayDriverJson = Record<string, GatewayStorageJson>;
+export type GatewayDriverJson = Record<string, GatewayJson>;
